apps/worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
boto3==1.29.7
minio==7.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
structlog==23.2.0
httpx==0.25.2
openai==1.3.7
anthropic==0.7.8
PyPDF2==3.0.1
python-docx==1.1.0
mlflow==2.8.1
pytest==7.4.3
pytest-asyncio==0.21.1
ruff==0.1.6
black==23.11.0
